requirements.txt
pyopengl
numpy
pygame
Pillow